service/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
confluent-kafka==2.6.0
idna==3.10
iter-tools==0.0.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-confluent-kafka==0.49b1
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp==1.28.1
